# docs/openapi.yaml
openapi: 3.0.0
info:
  title: AICleaner v3 Core Service
  version: "1.0.0"
  description: |
    A simplified, power-user-focused API for AI inference, configuration, and monitoring.
    This specification defines the core endpoints for the AICleaner v3 service,
    focusing on simplicity and performance while providing maximum flexibility.
servers:
  - url: http://localhost:8000/v1
security:
  - ApiKeyAuth: []
paths:
...
  /config:
    get:
      summary: Get Configuration
      description: Retrieves current service configuration with redacted secrets.
      operationId: getConfiguration
      responses:
        '200':
          description: Current configuration with secrets redacted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      summary: Update Configuration
      description: Updates the service configuration. Secrets are write-only.
      operationId: updateConfiguration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigRequest'
      responses:
        '200':
          description: Configuration updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    ConfigResponse:
      type: object
      properties:
        ai_provider:
          type: string
          description: Default AI provider.
          example: "gemini"
        providers:
          type: object
          additionalProperties:
            type: object
            properties:
              enabled:
                type: boolean
              api_key_configured:
                type: boolean
              models:
                type: array
                items:
                  type: string
        mqtt:
          type: object
          properties:
            host:
              type: string
            port:
              type: integer
            auto_discover:
              type: boolean
            discovery_prefix:
              type: string
        performance:
          type: object
          properties:
            cache_enabled:
              type: boolean
            metrics_retention_days:
              type: integer
    ConfigRequest:
      type: object
      properties:
        ai_provider:
          type: string
          description: Default AI provider.
        api_keys:
          type: object
          additionalProperties:
            type: string
          description: API keys for providers (write-only).
        mqtt:
          type: object
          properties:
            host:
              type: string
            port:
              type: integer
            username:
              type: string
            password:
              type: string
            auto_discover:
              type: boolean
            discovery_prefix:
              type: string
        performance:
          type: object
          properties:
            cache_enabled:
              type: boolean
            metrics_retention_days:
              type: integer
    Error:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        details:
          type: object
          additionalProperties: true
      required:
        - code
        - message
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 400
            message: "Invalid request payload"
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 500
            message: "An unexpected error occurred."